requirements.txt
pygame
gtts
wikipedia-api
